+++ vm/oops/memOop.hpp
@@ -29,6 +29,8 @@ OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISE
 
 // should clean up order of member functions  -Urs
 
+#include <cstdint>
+
 class memOopDesc: public oopDesc {
  protected:
   // instance variable
@@ -47,7 +49,7 @@ class memOopDesc: public oopDesc {
   friend memOop as_memOop(void* p);
 
   // conversion from memOop to memOopDesc*
-  memOopDesc* addr() const		{ return (memOopDesc*) (int(this) - Mem_Tag); }
+  memOopDesc* addr() const		{ return (memOopDesc*) (intptr_t(this) - Mem_Tag); }
 
   // space operations, is_old/new work w/o conversion to memOopDesc*
   // since oop > pointer (Mem_Tag >= 0)!
@@ -188,6 +190,6 @@ class memOopDesc: public oopDesc {
 
 inline memOop as_memOop(void* p)
 {
-    assert((int(p) & Tag_Mask) == 0, "not an aligned C pointer");
-    return memOop(int(p) + Mem_Tag);
+    assert((intptr_t(p) & Tag_Mask) == 0, "not an aligned C pointer");
+    return memOop(intptr_t(p) + Mem_Tag);
 }
